+++ concordium-std/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased changes
 
+- `concordium-std-derive`: add support for event schemas in the schema derivation macro.
 - `concordium-std-derive`: allow `#[concordium(state_parameter)]`'s value be not just identifier
   but any type path for `derive(DeserialWithState)` and `derive(Deletable)` to generate implementations.